+++ src/main/java/com/supermartijn642/packedup/BackpackInventory.java
@@ -9,7 +9,7 @@ import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.ShulkerBoxBlock;
 
-import java.io.File;
+import java.nio.file.Path;
 import java.util.*;
 
 /**
@@ -101,7 +101,7 @@ public class BackpackInventory {
         return stack1.isEmpty() || stack2.isEmpty() || (stack1.getItem() == stack2.getItem() && stack1.getDamageValue() == stack2.getDamageValue() && ItemStack.isSameItemSameTags(stack1, stack2));
     }
 
-    public void save(File file){
+    public void save(Path path){
         CompoundTag compound = new CompoundTag();
         compound.putInt("stacks", this.stacks.size());
         for(int slot = 0; slot < this.stacks.size(); slot++)
@@ -110,14 +110,14 @@ public class BackpackInventory {
         compound.putIntArray("bagsThisBagIsIn", Lists.newArrayList(this.bagsThisBagIsIn));
         compound.putInt("layer", this.layer);
         try{
-            NbtIo.write(compound, file);
+            NbtIo.write(compound, path);
         }catch(Exception e){e.printStackTrace();}
     }
 
-    public void load(File file){
+    public void load(Path path){
         CompoundTag compound;
         try{
-            compound = NbtIo.read(file);
+            compound = NbtIo.read(path);
         }catch(Exception e){
             e.printStackTrace();
             return;