+++ changelog.md
@@ -1,2 +1,2 @@
 ### Packed Up 1.0.30
-- Initial release of Packed Up for Minecraft 1.20
+- Initial release of Packed Up for Minecraft 1.20.3 & 1.20.4